Add gogrs cache
