--- cmd/twsecache/README.md
@@ -1,20 +1,0 @@
-gogrs - twsecache
-==================
-
-[![GoDoc](https://godoc.org/github.com/toomore/gogrs?status.svg)](https://godoc.org/github.com/toomore/gogrs/cmd/twsecache)
-[![Build Status](https://travis-ci.org/toomore/gogrs.svg?branch=master)](https://travis-ci.org/toomore/gogrs)
-
-清除 twse cache.
-
-Install:
-
-	go install github.com/toomore/gogrs/cmd/twsecache
-
-Usage:
-
-	twsecache [flags]
-
-The flags are:
-
-	-flushall
-		Clear cache.
